+++ package.json
@@ -29,7 +29,7 @@
     "test:smoke-vic": "cd examples/vic-gov-au/ && yarn test:smoke",
     "test:smoke-example": "cd examples/vic-gov-au/ && yarn test:smoke",
     "test:unit": "jest --colors",
-    "test:report": "node ./scripts/cucumberconvertor.js",
+    "test:report": "node ./packages/ripple-test-tools/scripts/cucumberconvertor.js ./test-results/cucumber",
     "build:storybook": "cd src && yarn build-storybook",
     "build:vic-gov-au": "cd examples/vic-gov-au && yarn run build",
     "lint": "sass-lint && yarn lint:es",
@@ -63,7 +63,6 @@
     "babel-preset-vue": "^2.0.2",
     "core-js": "^2.6.5",
     "cross-env": "^5.2.0",
-    "cucumber-junit-convert": "^1.1.0",
     "eslint": "^5.11.1",
     "eslint-config-standard": "^12.0.0",
     "eslint-friendly-formatter": "^3.0.0",
