+++ packages/ripple-test-tools/package.json
@@ -8,11 +8,16 @@
       "name": "Dylan Kelly <dylan.kelly@dpc.vic.gov.au>"
     }
   ],
+  "bin": {
+    "@dpc-sdp/ripple-test-tools/process-report": "/scripts/cucumberconvertor.js"
+  },
   "main": "index.js",
   "dependencies": {
     "csvtojson": "^2.0.8",
     "faker": "^4.1.0",
-    "puppeteer": "^1.20.0"
+    "puppeteer": "^1.20.0",
+    "glob-fs": "^0.1.7",
+    "cucumber-junit-convert": "^1.1.0"
   },
   "publishConfig": {
     "access": "public"
